+++ test/integrations/bronto.js
@@ -5,6 +5,7 @@ describe('Bronto', function(){
   var test = require('integration-tester');
   var analytics = require('analytics');
   var assert = require('assert');
+  var equal = require('equals');
   var sinon = require('sinon');
   var bronto;
 
@@ -82,6 +83,29 @@ describe('Bronto', function(){
         .track('some-event', { revenue: 50 })
         .called(bronto.bta.addConversionLegacy)
         .with('$', 'some-event', 50);
+    })
+  })
+
+  describe('ecommerce', function(){
+    beforeEach(function(){
+      bronto.bta = { addConversion: sinon.spy() };
+      bronto.initialize();
+    })
+
+    it('should send ecommerce data', function(){
+      var date = new Date;
+
+      test(bronto).track('completed order', {
+        products: [{ sku: 'c546c96', quantity: 8, name: 'my-product', price: 99.99 }],
+        orderId: '55c497bf',
+        date: date,
+      });
+
+      assert(equal(bronto.bta.addConversion.args[0][0], {
+        items: [{ item_id: 'c546c96', quantity: 8, desc: 'my-product', amount: 99.99 }],
+        order_id: '55c497bf',
+        date: date
+      }));
     })
   })