+++ test/integrations/foxmetrics.js
@@ -226,4 +226,93 @@ describe('FoxMetrics', function () {
     });
   });
 
+  describe('ecommerce', function(){
+    beforeEach(function(){
+      foxmetrics.initialize();
+      window._fxm.push = sinon.spy();
+    });
+
+    it('should track viewed product', function(){
+      test(foxmetrics).track('viewed product', {
+        sku: 'f84d349b',
+        name: 'my-product',
+        category: 'category'
+      });
+
+      assert(window._fxm.push.calledWith([
+        '_fxm.ecommerce.productview',
+        'f84d349b',
+        'my-product',
+        'category'
+      ]));
+    })
+
+    it('should track added product', function(){
+      test(foxmetrics).track('added product', {
+        id: 'c1ec1864',
+        name: 'my-product',
+        category: 'category'
+      });
+
+      assert(window._fxm.push.calledWith([
+        '_fxm.ecommerce.cartitem',
+        'c1ec1864',
+        'my-product',
+        'category'
+      ]));
+    })
+
+    it('should track removed product', function(){
+      test(foxmetrics).track('removed product', {
+        sku: 'c1ec1864',
+        name: 'my-product'
+      });
+
+      assert(window._fxm.push.calledWith([
+        '_fxm.ecommerce.removecartitem',
+        'c1ec1864',
+        'my-product',
+        undefined
+      ]));
+    })
+
+    it('should track completed order', function(){
+      test(foxmetrics).track('completed order', {
+        orderId: '3723ee8a',
+        total: 300,
+        tax: 10,
+        shipping: 20,
+        products: [{
+          sku: 'd370b4cd',
+          name: 'sony pulse',
+          category: 'tech',
+          price: 270,
+          quantity: 1
+        }]
+      });
+
+      assert(equal(window._fxm.push.args[0][0], [
+        '_fxm.ecommerce.order',
+        '3723ee8a',
+        270,
+        20,
+        10,
+        undefined,
+        undefined,
+        undefined,
+        1,
+      ]));
+
+      assert(equal(window._fxm.push.args[1][0], [
+        '_fxm.ecommerce.purchaseitem',
+        'd370b4cd',
+        'sony pulse',
+        'tech',
+        1,
+        270,
+        '3723ee8a'
+      ]));
+    })
+  })
+
 });