+++ test/integrations/gosquared.js
@@ -196,4 +196,39 @@ describe('GoSquared', function () {
     });
   });
 
+  describe('ecommerce', function(){
+    beforeEach(function(done){
+      gosquared.initialize();
+      gosquared.once('load', function(){
+        window._gs = sinon.spy();
+        done();
+      });
+    })
+
+    it('should send a transaction', function(){
+      test(gosquared)
+        .track('completed order', {
+          orderId: 'a9173991',
+          total: 90,
+          quantity: 10,
+          products: [{
+            category: 'my category',
+            name: 'my-product',
+            quantity: 10,
+            price: 9
+          }]
+        });
+
+      assert(window._gs.calledWith('transaction', 'a9173991', {
+        revenue: 90,
+        track: true
+      }, [{
+        name: 'my-product',
+        category: 'my category',
+        quantity: 10,
+        price: 9
+      }]));
+    })
+  })
+
 });